% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcf.R
\name{autoplot.mcf}
\alias{autoplot.mcf}
\title{Plot a mean cumulative function}
\usage{
\method{autoplot}{mcf}(object, ...)
}
\arguments{
\item{object}{an object of class \code{mcf}.}

\item{...}{ignored; kept for compatibility with the generic.}
}
\value{
A \code{ggplot} object.
}
\description{
Draws the mean cumulative function estimated by \code{\link[=mcf]{mcf()}} as a step
function, one colour per group.
}
\examples{
data(colon)
autoplot(mcf(Survr(hc, time, event) ~ as.factor(dukes), data = colon))
}
\seealso{
\code{\link[=mcf]{mcf()}}, \code{\link[=theme_survrec]{theme_survrec()}}
}
