% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDM3df.R
\name{LDM3df}
\alias{LDM3df}
\title{Landmark estimator for the general case of K gap times distribution function.}
\usage{
LDM3df(object, x, y, z)
}
\arguments{
\item{object}{An object of class multidf.}

\item{x}{The first time for obtaining estimates for the general case of distribution function.}

\item{y}{The second time for obtaining estimates for the general case of distribution function.}

\item{z}{The third time for obtaining estimates for the general case of distribution function.}
}
\description{
Provides estimates for the general case of K gap times distribution function based on landmarking.
}
\examples{
b4state <- multidf(time1=bladder5state$y1, event1=bladder5state$d1,
time2= bladder5state$y1+bladder5state$y2, event2=bladder5state$d2,
time=bladder5state$y1+bladder5state$y2+bladder5state$y3,
status=bladder5state$d3)
head(b4state)[[1]]

LDM3df(b4state,x=13,y=20,z=40)

b4 <- multidf(time1=bladder4$t1, event1=bladder4$d1,
             time2= bladder4$t2, event2=bladder4$d2,
             time=bladder4$t3, status=bladder4$d3)
LDM3df(b4,x=13,y=20,z=40)

}
\references{
van Houwelingen, H.C. (2007). Dynamic prediction by landmarking in event history analysis,
Scandinavian Journal of Statistics, 34, 70-85.
Kaplan, E. and Meier, P. (1958). Nonparametric Estimation from Incomplete Observations, Journal of the
American Statistical Association 53(282), 457-481.
}
\seealso{
\code{\link{LDM3df}}, \code{\link{LIN3df}} and \code{\link{WCH3df}}.
}
\author{
Gustavo Soutinho and Luis Meira-Machado
}
