% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcf.R
\name{mcf}
\alias{mcf}
\alias{print.mcf}
\title{Mean cumulative function of recurrent event data}
\usage{
mcf(formula, data)

\method{print}{mcf}(x, ...)
}
\arguments{
\item{formula}{a formula object with a \code{\link[=Survr]{Survr()}} object as the response
on the left of the \code{~} operator and a grouping term (or \code{1}) on the
right.}

\item{data}{a data frame in which to interpret the variables named in
the formula.}

\item{x}{an \code{mcf} object.}

\item{...}{further arguments passed to \code{\link[=print.data.frame]{print.data.frame()}}.}
}
\value{
An object of class \code{mcf}: a data frame with columns \code{group},
\code{time}, \code{n.event}, \code{n.risk} and \code{mcf}, with \code{print} and \code{autoplot}
methods.
}
\description{
Nonparametric (Nelson-Aalen type) estimate of the mean cumulative
function, i.e. the expected number of events experienced by a subject
up to each calendar time, optionally by group.
}
\details{
Calendar times are reconstructed as the within-subject cumulative sums
of the gap times of the \code{\link[=Survr]{Survr()}} response. At each distinct event
calendar time \eqn{t} the estimate increases by \eqn{dN(t)/Y(t)}, where
\eqn{dN(t)} is the number of events observed at \eqn{t} and \eqn{Y(t)}
the number of subjects still under observation (total observation time
\eqn{\ge t}).
}
\examples{
data(MMC)
m <- mcf(Survr(id, time, event) ~ group, data = MMC)
m
autoplot(m)
}
\references{
Lawless, J.F. and Nadeau, C. (1995). Some Simple Robust Methods for
the Analysis of Recurrent Events. \emph{Technometrics} \strong{37}, 158--168.
}
\seealso{
\code{\link[=autoplot.mcf]{autoplot.mcf()}}, \code{\link[=survfitr]{survfitr()}}
}
\keyword{survival}
