% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidf.R
\name{multidf}
\alias{multidf}
\title{Create a multidf object}
\usage{
multidf(time1, time, event1, status, ...)
}
\arguments{
\item{time1}{First time or censoring time.}

\item{time}{The total time of the process.}

\item{event1}{Indicator of the first time; 0 if the first time is censored and 1 otherwise.}

\item{status}{Censoring indicator of the survival time of the process; 0 if the total time is
censored and 1 otherwise.}

\item{...}{Other options. For instance, for three gap times: \code{time2} represent the second
time and \code{event2} the censoring indicator of the second time; 0 if the second time is censored and 1 otherwise. Other additional arguments, such as covariates, can also be included in the data set.}
}
\value{
An object of class "multidf". "multidf" objects are implemented as a single data frame.
}
\description{
Creates a "multidf" object, usually used as a response variable in a model formula.
}
\details{
Arguments in this function must be introduced in the following
order: \code{time1}, \code{event1}, \code{time} and \code{status}, where
\code{time1} and \code{time} are ordered event times and
\code{event1} and \code{event} their corresponding indicator statuses.
Other arguments can be also added. These should consider intermediate times and corresponding
censoring indicators or covariates.
}
\examples{
data("bladder4state")
b3state <- multidf(time1=bladder4state$y1, event1=bladder4state$d1, 
                  time=bladder4state$y1+bladder4state$y2,status=bladder4state$d2)
head(b3state[[1]])
data("bladder4")

b4 <- multidf(time1=bladder4$t1, event1=bladder4$d1,
             time2= bladder4$t2, event2=bladder4$d2,
             time=bladder4$t3, status=bladder4$d3)
head(b4state[[1]])

}
\author{
Gustavo Soutinho and Luis Meira-Machado
}
