% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NWW.R
\name{NWW}
\alias{NWW}
\title{Nadaraya-Watson weights}
\usage{
NWW(covariate, x, kernel = "gaussian", bw)
}
\arguments{
\item{covariate}{Covariate values for obtaining weights.}

\item{x}{Covariate value to compute the weight at.}

\item{kernel}{A character string specifying the desired kernel. See details below for possible options. 
Defaults to "gaussian" where the gaussian density kernel will be used.}

\item{bw}{A single numeric value to compute a kernel density bandwidth.}
}
\value{
A vector with Nadaraya-Watson weights.
}
\description{
Computes the Nadaraya-Watson weights.
}
\details{
Possible options for argument window are "gaussian", "epanechnikov", "tricube", "boxcar",
"triangular", "quartic" or "cosine".
}
\examples{
b3state2 <- multidf(time1=bladder4state$y1, event1=bladder4state$d1, 
time=bladder4state$y1+bladder4state$y2,status=bladder4state$d2, size=bladder4state$size)
obj0 <- b3state2[[1]]
NWW(covariate = obj0$size, x=3, kernel = "gaussian", bw = 3)

}
\author{
Gustavo Soutinho and Luis Meira-Machado
}
