% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.survfitr.R
\name{plot.survfitr}
\alias{plot.survfitr}
\alias{lines.survfitr}
\title{Plot survival curves of recurrent event data}
\usage{
\method{plot}{survfitr}(x, conf.int = TRUE, prob = FALSE, ...)

\method{lines}{survfitr}(x, prob = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{survfitr} (output of \code{\link[=survfitr]{survfitr()}},
\code{\link[=psh_fit]{psh_fit()}}, \code{\link[=wc_fit]{wc_fit()}} or \code{\link[=mlefrailty_fit]{mlefrailty_fit()}}).}

\item{conf.int}{if \code{TRUE} (default), pointwise 95\\% confidence bands
are drawn for the estimators that provide standard errors.}

\item{prob}{if \code{TRUE} the probability function (1 - survival) is drawn
instead of the survival function.}

\item{...}{additional arguments passed to \code{\link[=plot]{plot()}}.}
}
\value{
No return value; called for its side effect.
}
\description{
Plots the estimated survival (or probability) function from a
\code{survfitr} object, one curve per group when the fit has strata.
Additional fits can be added to the same axes with the \code{lines} method.
}
\examples{
data(MMC)
fit <- survfitr(Survr(id, time, event) ~ group,
  data = MMC,
  type = "wang-chang"
)
plot(fit)
}
\seealso{
\code{\link[=psh_fit]{psh_fit()}}, \code{\link[=wc_fit]{wc_fit()}}, \code{\link[=mlefrailty_fit]{mlefrailty_fit()}}
}
\keyword{survival}
