% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survrec-package.R
\docType{package}
\name{survrec-package}
\alias{survrec}
\alias{survrec-package}
\title{survrec: Survival Analysis for Recurrent Event Data}
\description{
Estimation of the survival function of inter-occurrence times for
recurrent event data. Three estimators are available: the generalized
product-limit estimator of Peña, Strawderman and Hollander (2001), the
estimator of Wang and Chang (1999) for correlated inter-occurrence
times, and maximum likelihood estimation under a gamma frailty model.
Survival quantiles can be compared between groups through several
bootstrap schemes.
}
\details{
The main entry points are \code{\link[=Survr]{Survr()}} to build the response object,
\code{\link[=survfitr]{survfitr()}} to estimate survival curves (optionally by group) and
\code{\link[=survdiffr]{survdiffr()}} to bootstrap survival quantiles.
}
\references{
Peña, E.A., Strawderman, R. and Hollander, M. (2001). Nonparametric
Estimation with Recurrent Event Data. \emph{Journal of the American
Statistical Association} \strong{96}, 1299--1315.

Wang, M.-C. and Chang, S.-H. (1999). Nonparametric Estimation of a
Recurrent Survival Function. \emph{Journal of the American Statistical
Association} \strong{94}, 146--153.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/isglobal-brge/survrec}
  \item Report bugs at \url{https://github.com/isglobal-brge/survrec/issues}
}

}
\author{
\strong{Maintainer}: Dolors Pelegri-Siso \email{dolors.pelegri@isglobal.org} (\href{https://orcid.org/0000-0002-5993-3003}{ORCID})

Authors:
\itemize{
  \item Juan R. González \email{juanr.gonzalez@isglobal.org}
  \item Edsel A. Peña
  \item Robert L. Strawderman
}

}
