% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_fit.R
\name{wc_fit}
\alias{wc_fit}
\title{Wang-Chang survival estimator}
\usage{
wc_fit(x, tvals)
}
\arguments{
\item{x}{a survival recurrent event object, see \code{\link[=Survr]{Survr()}}.}

\item{tvals}{optional vector of times at which the survival function is
also evaluated.}
}
\value{
A list with class \code{survfitr} containing:
\describe{
\item{n}{number of units or subjects observed.}
\item{m}{number of events of each subject.}
\item{failed}{uncensored gap times, ordered by subject.}
\item{censored}{censored gap time of each subject.}
\item{time}{ordered distinct event times.}
\item{n.event}{weighted number of events at each distinct time.}
\item{AtRisk}{weighted number of units at risk at each distinct
time.}
\item{survfunc}{survival estimate at each distinct time.}
\item{std.error}{standard error of the survival estimate.}
\item{tvals}{copy of the \code{tvals} argument.}
\item{WCpleAttvals}{survival estimate at the \code{tvals} times.}
}
}
\description{
Estimates the survival function of correlated or i.i.d. inter-occurrence
times of recurrent event data with the product-limit estimator of Wang
and Chang (1999).
}
\details{
Wang and Chang (1999) proposed an estimator of the common marginal
survivor function for the case where within-unit inter-occurrence times
are correlated. The correlation structure they considered is quite
general and contains both the i.i.d. and the multiplicative (hence
gamma) frailty model as special cases.

This estimator removes the bias noted for the product-limit estimator of
Peña, Strawderman and Hollander (2001) when inter-occurrence times are
correlated within units. However, when applied to i.i.d.
inter-occurrence times it is not expected to perform as well as the PSH
estimator, especially with regard to efficiency.
}
\note{
The maintainers wish to thank Professors Chiung-Yu Huang and
Shu-Hui Chang for providing the original Fortran code computing the
standard errors of Wang and Chang's estimator.
}
\examples{
data(MMC)
fit <- wc_fit(Survr(MMC$id, MMC$time, MMC$event))
fit
plot(fit, conf.int = FALSE)

# compare with Peña-Strawderman-Hollander
fit <- psh_fit(Survr(MMC$id, MMC$time, MMC$event))
lines(fit, lty = 2)

# and with the gamma frailty MLE
fit <- mlefrailty_fit(Survr(MMC$id, MMC$time, MMC$event))
lines(fit, lty = 3)
}
\references{
Wang, M.-C. and Chang, S.-H. (1999). Nonparametric Estimation of a
Recurrent Survival Function. \emph{Journal of the American Statistical
Association} \strong{94}, 146--153.
}
\seealso{
\code{\link[=survfitr]{survfitr()}}, \code{\link[=Survr]{Survr()}}
}
\keyword{survival}
