% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{survrec-deprecated}
\alias{survrec-deprecated}
\alias{psh.fit}
\alias{wc.fit}
\alias{mlefrailty.fit}
\alias{surv.search}
\alias{q.search}
\title{Deprecated functions in survrec}
\usage{
psh.fit(x, tvals)

wc.fit(x, tvals)

mlefrailty.fit(
  x,
  tvals,
  lambda = NULL,
  alpha = NULL,
  alpha.min,
  alpha.max,
  tol = 1e-07,
  maxiter = 500,
  alpha.console = TRUE
)

surv.search(tvals, time, surv)

q.search(f, q = 0.5)
}
\arguments{
\item{x, tvals, lambda, alpha, alpha.min, alpha.max, tol, maxiter, alpha.console}{see the replacement functions.}

\item{time, surv, f, q}{see the replacement functions.}
}
\description{
The dot-separated names of survrec 1.x were renamed to snake_case in
version 2.0.0. The old names still work but emit a deprecation warning
and will be removed in a future release.
}
\details{
\describe{
\item{\code{psh.fit()}}{use \code{\link[=psh_fit]{psh_fit()}} instead.}
\item{\code{wc.fit()}}{use \code{\link[=wc_fit]{wc_fit()}} instead.}
\item{\code{mlefrailty.fit()}}{use \code{\link[=mlefrailty_fit]{mlefrailty_fit()}} instead.}
\item{\code{surv.search()}}{use \code{\link[=surv_search]{surv_search()}} instead.}
\item{\code{q.search()}}{use \code{\link[=q_search]{q_search()}} instead.}
}
}
\keyword{internal}
